+++ ext/duckdb/ruby-duckdb.h
@@ -50,6 +50,7 @@
 #include "./memory_helper.h"
 #include "./table_function.h"
 #include "./table_description.h"
+#include "./arrow_array_stream.h"
 
 extern VALUE mDuckDB;
 extern VALUE cDuckDBDatabase;